+++ modules/queue/src/queue.cpp
@@ -2,7 +2,7 @@
 
 #include "include/queue.h"
 
-Queue::Queue(size_t s) {
+Queue::Queue(int s) {
     if (s < 0)
         throw "Queue size < 0";
     tail_ = -1;
